forex-bot/requirements.txt
oandapyV20==0.7.2
python-dotenv>=1.0.1
click>=8.1.7
requests>=2.32.3
six>=1.16.0
